Made some adjustments to prepare for gas reduction

+++ contracts/CawClientManager.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.8.0;
 
 struct CawClient {
-  uint64 id;
+  uint32 id;
   address feeAddress;
   address ownerAddress;
   uint256 withdrawFee;
@@ -13,16 +13,16 @@ struct CawClient {
 contract CawClientManager {
 
   address buyAndBurnAddress;
-  uint64 public nextClientId = 1;
-  mapping(uint64 => CawClient) public clients;
+  uint32 public nextClientId = 1;
+  mapping(uint32 => CawClient) public clients;
 
-  event ClientCreated(uint64 nextClientId, CawClient client);
+  event ClientCreated(uint32 nextClientId, CawClient client);
 
   constructor(address _buyAndBurn) {
     buyAndBurnAddress = _buyAndBurn;
   }
 
-  modifier onlyClientOwner(uint64 clientId) {
+  modifier onlyClientOwner(uint32 clientId) {
     require(clients[clientId].ownerAddress == msg.sender, "Not the owner");
     _;
   }
@@ -32,46 +32,46 @@ contract CawClientManager {
    * @param clientId The ID of the client.
    * @return The CawClient struct.
    */
-  function getClient(uint64 clientId) public view returns (CawClient memory) {
+  function getClient(uint32 clientId) public view returns (CawClient memory) {
     return clients[clientId];
   }
 
-  function getMintFee(uint64 clientId) public view returns (uint256) {
+  function getMintFee(uint32 clientId) public view returns (uint256) {
     CawClient storage client = clients[clientId];
     return client.mintFee;
   }
 
-  function getAuthFee(uint64 clientId) public view returns (uint256) {
+  function getAuthFee(uint32 clientId) public view returns (uint256) {
     CawClient storage client = clients[clientId];
     return client.authFee;
   }
 
-  function getDepositFee(uint64 clientId) public view returns (uint256) {
+  function getDepositFee(uint32 clientId) public view returns (uint256) {
     CawClient storage client = clients[clientId];
     return client.depositFee;
   }
 
-  function getWithdrawFee(uint64 clientId) public view returns (uint256) {
+  function getWithdrawFee(uint32 clientId) public view returns (uint256) {
     CawClient storage client = clients[clientId];
     return client.withdrawFee;
   }
 
-  function getMintFeeAndAddress(uint64 clientId) public view returns (uint256, address) {
+  function getMintFeeAndAddress(uint32 clientId) public view returns (uint256, address) {
     CawClient storage client = clients[clientId];
     return (client.mintFee, client.feeAddress);
   }
 
-  function getAuthFeeAndAddress(uint64 clientId) public view returns (uint256, address) {
+  function getAuthFeeAndAddress(uint32 clientId) public view returns (uint256, address) {
     CawClient storage client = clients[clientId];
     return (client.authFee, client.feeAddress);
   }
 
-  function getDepositFeeAndAddress(uint64 clientId) public view returns (uint256, address) {
+  function getDepositFeeAndAddress(uint32 clientId) public view returns (uint256, address) {
     CawClient storage client = clients[clientId];
     return (client.depositFee, client.feeAddress);
   }
 
-  function getWithdrawFeeAndAddress(uint64 clientId) public view returns (uint256, address) {
+  function getWithdrawFeeAndAddress(uint32 clientId) public view returns (uint256, address) {
     CawClient storage client = clients[clientId];
     return (client.withdrawFee, client.feeAddress);
   }
@@ -101,7 +101,7 @@ contract CawClientManager {
   * @param clientId The ID of the client.
     * @param newOwner The address of the new owner.
     */
-  function changeOwner(uint64 clientId, address newOwner) public onlyClientOwner(clientId) {
+  function changeOwner(uint32 clientId, address newOwner) public onlyClientOwner(clientId) {
     clients[clientId].ownerAddress = newOwner;
   }
 
@@ -110,7 +110,7 @@ contract CawClientManager {
   * @param clientId The ID of the client.
     * @param fee The new withdraw fee.
       */
-  function setWithdrawFee(uint64 clientId, uint256 fee) public onlyClientOwner(clientId) {
+  function setWithdrawFee(uint32 clientId, uint256 fee) public onlyClientOwner(clientId) {
     clients[clientId].withdrawFee = fee;
   }
 
@@ -119,7 +119,7 @@ contract CawClientManager {
    * @param clientId The ID of the client.
    * @param fee The new auth fee.
    */
-  function setAuthFee(uint64 clientId, uint256 fee) public onlyClientOwner(clientId) {
+  function setAuthFee(uint32 clientId, uint256 fee) public onlyClientOwner(clientId) {
     clients[clientId].authFee = fee;
   }
 
@@ -128,7 +128,7 @@ contract CawClientManager {
    * @param clientId The ID of the client.
    * @param fee The new deposit fee.
    */
-  function setDepositFee(uint64 clientId, uint256 fee) public onlyClientOwner(clientId) {
+  function setDepositFee(uint32 clientId, uint256 fee) public onlyClientOwner(clientId) {
     clients[clientId].depositFee = fee;
   }
 
@@ -137,7 +137,7 @@ contract CawClientManager {
    * @param clientId The ID of the client.
    * @param fee The new mint fee.
    */
-  function setMintFee(uint64 clientId, uint256 fee) public onlyClientOwner(clientId) {
+  function setMintFee(uint32 clientId, uint256 fee) public onlyClientOwner(clientId) {
     clients[clientId].mintFee = fee;
   }
 
@@ -146,7 +146,7 @@ contract CawClientManager {
    * @param clientId The ID of the client.
    * @param feeAddress The new fee address.
    */
-  function setFeeAddress(uint64 clientId, address feeAddress) public onlyClientOwner(clientId) {
+  function setFeeAddress(uint32 clientId, address feeAddress) public onlyClientOwner(clientId) {
     clients[clientId].feeAddress = feeAddress;
   }
 }
