+++ contracts/CawNameMinter.sol
@@ -8,7 +8,7 @@ import "./interfaces/IMint.sol";
 
 contract CawNameMinter is Context {
 
-  mapping(string => uint64) public idByUsername;
+  mapping(string => uint32) public idByUsername;
 
   IMint CawName;
   IERC20 CAW;
@@ -18,7 +18,7 @@ contract CawNameMinter is Context {
     CawName = IMint(_cawNames);
   }
 
-  function mint(uint64 clientId, string memory username, uint256 lzTokenAmount) public payable {
+  function mint(uint32 clientId, string memory username, uint256 lzTokenAmount) public payable {
     require(idByUsername[username] == 0, "Username has already been taken");
     require(isValidUsername(username), "Username must only consist of 1-255 lowercase letters and numbers");
     uint256 amount = costOfName(username);
@@ -27,7 +27,7 @@ contract CawNameMinter is Context {
     require(CAW.allowance(_msgSender(), address(this)) >= amount, "You must approve spending of your CAW");
     CAW.transferFrom(_msgSender(), address(0xdEAD000000000000000042069420694206942069), amount);
 
-    uint64 newId = CawName.nextId();
+    uint32 newId = CawName.nextId();
     idByUsername[username] = newId;
 
     CawName.mint{value: msg.value}(clientId, msg.sender, username, newId, lzTokenAmount);
